[80] Refactor code
