+++ domain/src/main/java/co/nimblehq/compose/crypto/domain/model/CoinItem.kt
@@ -40,3 +40,4 @@ data class CoinItem(
 }
 
 
+
